WIP separate pivot and compass rose (not compiling)
