+++ editor/src/consts.rs
@@ -54,16 +54,17 @@ pub const DEFAULT_STROKE_WIDTH: f64 = 2.;
 pub const SELECTION_TOLERANCE: f64 = 5.;
 pub const DRAG_DIRECTION_MODE_DETERMINATION_THRESHOLD: f64 = 15.;
 pub const SELECTION_DRAG_ANGLE: f64 = 90.;
+
+// PIVOT
 pub const PIVOT_CROSSHAIR_THICKNESS: f64 = 1.;
 pub const PIVOT_CROSSHAIR_LENGTH: f64 = 9.;
+pub const PIVOT_DIAMETER: f64 = 5.;
 
+// COMPASS ROSE
 pub const COMPASS_ROSE_ANGLE_WIDTH: f64 = 20.; // Must be less than 45
-
-pub const COMPASS_ROSE_PIVOT_DIAMETER: f64 = 5.;
 pub const COMPASS_ROSE_RING_INNER_DIAMETER: f64 = 13.;
 pub const COMPASS_ROSE_MAIN_RING_DIAMETER: f64 = 15.;
 pub const COMPASS_ROSE_HOVER_RING_DIAMETER: f64 = 23.;
-
 pub const COMPASS_ROSE_ARROW_SIZE: f64 = 5.;
 
 // TRANSFORM OVERLAY
