+++ editor/src/messages/tool/common_functionality/mod.rs
@@ -1,5 +1,6 @@
 pub mod auto_panning;
 pub mod color_selector;
+pub mod compass_rose;
 pub mod graph_modification_utils;
 pub mod measure;
 pub mod pivot;
