+++ editor/src/messages/tool/common_functionality/pivot.rs
@@ -1,7 +1,7 @@
 //! Handler for the pivot overlay visible on the selected layer(s) whilst using the Select tool which controls the center of rotation/scale and origin of the layer.
 
 use super::graph_modification_utils;
-use crate::consts::{COMPASS_ROSE_ANGLE_WIDTH, COMPASS_ROSE_HOVER_RING_DIAMETER, COMPASS_ROSE_PIVOT_DIAMETER, COMPASS_ROSE_RING_INNER_DIAMETER};
+use crate::consts::PIVOT_DIAMETER;
 use crate::messages::layout::utility_types::widget_prelude::*;
 use crate::messages::portfolio::document::overlays::utility_types::OverlayContext;
 use crate::messages::portfolio::document::utility_types::document_metadata::LayerNodeIdentifier;
@@ -9,53 +9,6 @@ use crate::messages::prelude::*;
 
 use glam::{DAffine2, DVec2};
 use std::collections::VecDeque;
-use std::f64::consts::FRAC_PI_2;
-
-#[derive(Clone, Debug, PartialEq)]
-pub enum CompassRoseState {
-	Pivot,
-	Ring,
-	AxisX,
-	AxisY,
-	None,
-}
-
-#[derive(Clone, Copy, Debug, PartialEq, Eq, Hash, Default)]
-pub enum Axis {
-	#[default]
-	None,
-	X,
-	Y,
-}
-
-impl Axis {
-	pub fn is_constraint(&self) -> bool {
-		matches!(self, Self::X | Self::Y)
-	}
-}
-
-impl CompassRoseState {
-	pub fn can_grab(&self) -> bool {
-		matches!(self, Self::Ring | Self::AxisX | Self::AxisY)
-	}
-
-	pub fn is_pivot(&self) -> bool {
-		matches!(self, Self::Pivot)
-	}
-
-	pub fn is_ring(&self) -> bool {
-		matches!(self, Self::Ring)
-	}
-
-	pub fn axis_type(&self) -> Option<Axis> {
-		match self {
-			CompassRoseState::AxisX => Some(Axis::X),
-			CompassRoseState::AxisY => Some(Axis::Y),
-			CompassRoseState::Ring => Some(Axis::None),
-			_ => None,
-		}
-	}
-}
 
 #[derive(Clone, Debug)]
 pub struct Pivot {
@@ -88,10 +41,6 @@ impl Pivot {
 		let bounds_transform = DAffine2::from_translation(min) * DAffine2::from_scale(max - min);
 		let layer_transform = document.metadata().transform_to_viewport(layer);
 		layer_transform * bounds_transform
-	}
-
-	pub fn get_compass_position(&self) -> DVec2 {
-		self.transform_from_normalized.transform_point2(DVec2::splat(0.5))
 	}
 
 	/// Recomputes the pivot position and transform.
@@ -134,11 +83,10 @@ impl Pivot {
 		}
 	}
 
-	pub fn update_pivot(&mut self, document: &DocumentMessageHandler, overlay_context: &mut OverlayContext, angle: f64, show_compass_with_hover_ring: Option<bool>) {
+	pub fn update_pivot(&mut self, document: &DocumentMessageHandler, overlay_context: &mut OverlayContext, angle: f64) {
 		self.recalculate_pivot(document);
 		if let Some(pivot) = self.pivot {
-			let compass_center = self.get_compass_position();
-			overlay_context.pivot(pivot, compass_center, angle, show_compass_with_hover_ring);
+			overlay_context.pivot(pivot, angle);
 		}
 	}
 
@@ -177,35 +125,7 @@ impl Pivot {
 	}
 
 	/// Answers if the pointer is currently positioned over the pivot.
-	pub fn compass_rose_state(&self, mouse: DVec2, angle: f64) -> CompassRoseState {
-		match self.pivot {
-			None => CompassRoseState::None,
-			Some(pivot) => {
-				let compass_center = self.get_compass_position();
-
-				let compass_distance_squared = mouse.distance_squared(compass_center);
-				let pivot_distance_squared = mouse.distance_squared(pivot);
-
-				if pivot_distance_squared < (COMPASS_ROSE_PIVOT_DIAMETER / 2.).powi(2) {
-					return CompassRoseState::Pivot;
-				}
-
-				if (COMPASS_ROSE_RING_INNER_DIAMETER / 2.).powi(2) < compass_distance_squared && compass_distance_squared < (COMPASS_ROSE_HOVER_RING_DIAMETER / 2.).powi(2) {
-					let angle = (mouse - compass_center).angle_to(DVec2::from_angle(angle)).abs();
-					let resolved_angle = (FRAC_PI_2 - angle).abs();
-					let width = COMPASS_ROSE_ANGLE_WIDTH.to_radians();
-
-					if resolved_angle < width {
-						CompassRoseState::AxisY
-					} else if resolved_angle > (FRAC_PI_2 - width) {
-						CompassRoseState::AxisX
-					} else {
-						CompassRoseState::Ring
-					}
-				} else {
-					CompassRoseState::None
-				}
-			}
-		}
+	pub fn is_over(&self, mouse: DVec2) -> bool {
+		self.pivot.filter(|&pivot| mouse.distance_squared(pivot) < (PIVOT_DIAMETER / 2.).powi(2)).is_some()
 	}
 }
